\name{adductsTable}
\alias{adductsTable}
\docType{data}
\title{
Adducts table
}
\description{
Table of possible adducts to be employed by LipidMS and related information.
}
\usage{data("adductsTable")}
\format{
  Data frame with 18 observations and the following 4 variables.
  \describe{
    \item{\code{adduct}}{character vector with the adducts names.}
    \item{\code{mdiff}}{numeric vector indicating the mass differences.}
    \item{\code{charge}}{numeric vector indicating the charge.}
    \item{\code{n}}{numeric vector. It indicates if the ion is a monomer (1),
    a dimer (2), etc.}
  }
}
\keyword{datasets}
