% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignDB.R
\name{assignDB}
\alias{assignDB}
\title{load LipidMS default data bases}
\usage{
assignDB()
}
\value{
list of data frames
}
\description{
load all LipidMS default data bases required to run identification functions.
}
\examples{
\donttest{
dbs <- assignDB()
}

}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
