\name{baconjdb}
\alias{baconjdb}
\docType{data}
\title{
Bile acids conjugates database
}
\description{
Common bile acids conjugates. It can be modified to look for other BA species.
}
\usage{data("baconjdb")}
\format{
  Data frame with 2 observations and the following 2 variables.
  \describe{
    \item{\code{total}}{character vector indicating the names of the conjugates.}
    \item{\code{Mass}}{numeric vector with the neutral masses of the conjugates
    fragments.}
  }
}
\keyword{datasets}
