\name{badb}
\alias{badb}
\docType{data}
\title{
Bile acids database
}
\description{
In silico generated database for common bile acids.
}
\usage{data("badb")}
\format{
  Data frame with 9 observations and the following 5 variables.
  \describe{
    \item{\code{formula}}{character vector with the molecular formulas.}
    \item{\code{total}}{character vector containing the names of the BAs (i.e.
    CA, TDCA, GLCA...).}
    \item{\code{Mass}}{numeric vector with the neutral masses.}
    \item{\code{conjugate}}{character vector containing the conjugate of each BA.}
    \item{\code{base}}{character vector containing the core of each BA.}
  }
}
\keyword{datasets}
