\name{cldb}
\alias{cldb}
\docType{data}
\title{
Cardiolipins database
}
\description{
In silico generated database for commo CLs.
}
\usage{data("cldb")}
\format{
  Data frame with 714 observations and the following 3 variables.
  \describe{
    \item{\code{formula}}{character vector containing molecular formulas.}
    \item{\code{total}}{character vector indicating the total number of carbons
    and double bounds of the chains.}
    \item{\code{Mass}}{numeric vector with the neutral masses.}
  }
}
\keyword{datasets}
