% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalProcessing.R
\name{clust}
\alias{clust}
\title{Clustering for MS peaks based on mz or RT.}
\usage{
clust(values, mins, maxs, samples, unique.samples, maxdist, ppm)
}
\arguments{
\item{values}{values to clusterize (mz or RT).}

\item{mins}{lower bound for each value.}

\item{maxs}{higher bound for each value.}

\item{samples}{numeric vector that indicates to which sample each value belongs.}

\item{unique.samples}{logical. FALSE if several measures from the same sample 
can be clusterized together.}

\item{maxdist}{maximum distance allowed within a cluster.}

\item{ppm}{logical. TRUE if maxdist is given in ppm.}
}
\value{
numeric vector with the assigned clusters
}
\description{
Clustering for MS peaks based on mz or RT.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
