% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{combineClassesAdducts}
\alias{combineClassesAdducts}
\title{Creates a list that maps lipid classes with their database name and default adducts.}
\usage{
combineClassesAdducts()
}
\value{
db_mapping, a list with Classes and Adducts
}
\description{
Creates a list that maps lipid classes with their database name and default adducts.
annotations and retention time predictions.
}
\author{
Celia Regal Marques <celia_regal@iislafe.es>
}
\keyword{internal}
