\name{confLevels}
\alias{confLevels}
\docType{data}
\title{
Confidence Annotation Levels
}
\description{
Confidence annotation levels and their hierarchy.
}
\usage{data("confLevels")}
\format{
  Data frame with 5 observations and 2 variables.
  \describe{
    \item{\code{level}}{character vector with the names of the annotation levels.}
    \item{\code{order}}{numeric vector that indicates the hierarchichal order.}
  }
}
\keyword{datasets}
