% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{dbsAdducts}
\alias{dbsAdducts}
\title{Create tables containing the m/z of lipids along with their preferred adducts}
\usage{
dbsAdducts(file = NULL, sep = ";", sep_adducts = ",", polarity, classes = NULL)
}
\arguments{
\item{file}{path csv with classes to predict}

\item{sep}{Character. Field separator used when reading the CSV file.}

\item{sep_adducts}{Character. Separator used to split multiple adducts within
the "Adducts" column of the file.}

\item{polarity}{character value: 'negative', 'positive'.}

\item{classes}{A vector containing the lipid classes for which retention time
is to be predicted. If classes = NULL, predictions will be made for all classes
where there are candidates and data to create a model.}
}
\value{
dbs_modif
}
\description{
Create tables containing the m/z of lipids along with their preferred adducts
for each polarity
}
\author{
Celia Regal Marques <celia_regal@iislafe.es>
}
\keyword{internal}
