% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{dbTwoChains}
\alias{dbTwoChains}
\title{Creation of a database for FAHFA, DG and PL.}
\usage{
dbTwoChains(chains, lipid)
}
\arguments{
\item{chains}{character vector indicating the FAs to be employed}

\item{lipid}{character value indication the class of lipid.}
}
\value{
data frame containing formula, mass and total number of carbons and
insaturations
}
\description{
Creation of a database for FAHFA, DG and PL.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
