% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{diffcb}
\alias{diffcb}
\title{Difference between two carbon:bounds structures}
\usage{
diffcb(total, frag, db)
}
\arguments{
\item{total}{character value indicating the precursor structure}

\item{frag}{character value indicating the fragment structure}

\item{db}{db of chains to be considered}
}
\value{
Character value
}
\description{
This function calculates the number of carbon and double bounds that differ
between two structures.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
