% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.with_future_plan}
\alias{.with_future_plan}
\title{Temporarily set a future plan and globals max size while running a block}
\usage{
.with_future_plan(workers = 1L, globals_gb = 24, expr)
}
\arguments{
\item{workers}{ncores}

\item{globals_gb}{max size for globals}

\item{expr}{expr}
}
\description{
Internal helper to encapsulate future::plan(...) and future.globals.maxSize
for a given expression, restoring the user's state on exit.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
