% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{findMS2precursor}
\alias{findMS2precursor}
\title{find lisnks between MS1 peaks and precursors selected for MS2 in DDA}
\usage{
findMS2precursor(mz, minrt, maxrt, precursors, ppm)
}
\arguments{
\item{mz}{mz}

\item{minrt}{minimum intensity}

\item{maxrt}{maximum intensity}

\item{precursors}{data frame with all precursors}

\item{ppm}{mass tolerance}
}
\value{
peak-pick based on previous EIC clusters generated by \link{clustering}
}
\description{
Find lisnks between MS1 peaks and precursors selected for MS2 in DDA.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
