% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalProcessing.R
\name{getallpeaks}
\alias{getallpeaks}
\title{Extract peaks from all msobjects in a msbatch.}
\usage{
getallpeaks(msbatch)
}
\arguments{
\item{msbatch}{msbatch object}
}
\value{
data.frame with 6 columns (mz, RT, int, peakID, isotope and isoGroup).
}
\description{
Extract peaks from all MS1 peaklists of the msobjects in a msbatch.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
