% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalProcessing.R
\name{getfeaturestable}
\alias{getfeaturestable}
\title{Write features table based on groups}
\usage{
getfeaturestable(msbatch)
}
\arguments{
\item{msbatch}{msbatch object}
}
\value{
data.frame
}
\description{
Write features table based on groups
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
