% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idFunctionsPos.R
\name{idAcylCerpos}
\alias{idAcylCerpos}
\title{Acylceramides (AcylCer) annotation for ESI+}
\usage{
idAcylCerpos(
  msobject,
  ppm_precursor = 5,
  ppm_products = 10,
  rttol = 3,
  rt,
  adducts = c("M+H", "M+H-H2O", "M+Na"),
  clfrags = c(),
  clrequired = c(),
  ftype = c(),
  chainfrags_sn1 = c("cbdiff-cer_M+H", "cbdiff-cer_M+H-H2O", "cbdiff-cer_M+H-2H2O"),
  chainfrags_sn2 = c("sph_M+H-H2O", "sph_M+H-2H2O"),
  chainfrags_sn3 = c("fa_Mn+0.02329"),
  intrules = c("sph_sn2/cbdiff-cer_sn1", "sph_sn2/fa_sn3"),
  rates = c("2/1", "5/1"),
  intrequired = c(T, T),
  coelCutoff = 0.8,
  dbs,
  verbose = TRUE
)
}
\arguments{
\item{msobject}{an msobject returned by \link{dataProcessing}.}

\item{ppm_precursor}{mass tolerance for precursor ions. By default, 5 ppm.}

\item{ppm_products}{mass tolerance for product ions. By default, 10 ppm.}

\item{rttol}{total rt window for coelution between precursor and product
ions. By default, 3 seconds.}

\item{rt}{rt range where the function will look for candidates. By default,
it will search within all RT range in MS1.}

\item{adducts}{expected adducts for Cer in ESI-. Adducts allowed can
be modified in adductsTable (dbs argument).}

\item{clfrags}{vector containing the expected fragments for a given lipid
class. See \link{checkClass} for details.}

\item{clrequired}{logical vector indicating if each class fragment is
required or not. If any of them is required, at least one of them must be
present within the coeluting fragments. See \link{checkClass} for details.}

\item{ftype}{character vector indicating the type of fragments in clfrags.
It can be: "F" (fragment), "NL" (neutral loss) or "BB" (building block).
See \link{checkClass} for details.}

\item{chainfrags_sn1}{character vector containing the fragmentation rules for
the sphingoid base. See \link{chainFrags} for details.}

\item{chainfrags_sn2}{character vector containing the fragmentation rules for
the chain fragments in sn2 position. See \link{chainFrags} for details. If
empty, it will be estimated based on the difference between precursors and
sn1 chains.}

\item{chainfrags_sn3}{character vector containing the fragmentation rules for
the acyl chain. See \link{chainFrags} for details.}

\item{intrules}{character vector specifying the fragments to compare. See
\link{checkIntensityRules}.}

\item{rates}{character vector with the expected ratesbetween fragments given
as a string (e.g. "3/1"). See \link{checkIntensityRules}.}

\item{intrequired}{logical vector indicating if any of the rules is required.
If not, at least one must be verified to confirm the structure.}

\item{coelCutoff}{coelution score threshold between parent and fragment ions.
Only applied if rawData info is supplied. By default, 0.8.}

\item{dbs}{list of data bases required for annotation. By default, dbs
contains the required data frames based on the default fragmentation rules.
If these rules are modified, dbs may need to be supplied. See \link{createLipidDB}
and \link{assignDB}.}

\item{verbose}{print information messages.}
}
\value{
annotated msobject (list with several elements). The results element
is a data frame that shows: ID, lipid class, CDB (total number of carbons
and double bounds), FA composition (specific chains composition if it has
been confirmed), mz, RT (in seconds), I (intensity), Adducts, ppm (mz error),
confidenceLevel (Subclass, FA level, where chains are known but not their
positions, or FA position level), peakID, and Score (parent-fragment coelution 
score mean in DIA data or relative sum intensity in DDA of all fragments used 
for the identification).
}
\description{
AcylCer identification based on fragmentation patterns for LC-MS/MS DIA or DDA
data acquired in positive mode.
}
\details{
\code{idAcylCerpos} function involves 5 steps. 1) FullMS-based
identification of candidate AcylCer as M+H, M+H-H2O and M+Na. 2) Search of 
AcylCer class fragments: there are no class fragments by default. 3) Search of 
specific fragments that inform about the acyl chain (Cer as M+H, M+H-H2O or
M+H-2H2H), the sphingoid base (Sph as M+H-H2O or M+H-2H2O) and the FA chain 
(FA as M+H but with a N intead of an O, what results in a mass difference of 
0.02329 with the Mn of the FA chain). 4) Look for possible chains structure 
based on the combination of chain fragments. 5) Check intensity rules to 
confirm chains position. In this case, Sph fragment must be twice more
intense than the loss of the acyl chain and at least 5 times more intense than
the FA chain from sn3.

Results data frame shows: ID, lipid class, CDB (total number
of carbons and double bounds), FA composition (specific chains composition if
it has been confirmed), mz, RT (in seconds), I (intensity, which comes
directly from de input), Adducts, ppm (mz error), confidenceLevel (Subclass,
FA level, where chains are known but not their positions, or FA position
level) and Score (parent-fragment coelution score mean in DIA data or relative 
sum intensity in DDA of all fragments used for the identification).
}
\note{
This function has been writen based on fragmentation patterns
observed for three different platforms (QTOF 6550 from Agilent, Synapt G2-Si
from Waters and Q-exactive from Thermo), but it may need to be customized for
other platforms or acquisition settings.
}
\examples{
\dontrun{
msobject <- idCerPneg(msobject)
}

}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
