% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idPEpos.R
\name{idPEpos}
\alias{idPEpos}
\title{Phosphoethanolamines (PE) annotation for ESI+}
\usage{
idPEpos(MS1, MSMS1, MSMS2 = data.frame(), ppm_precursor = 5,
  ppm_products = 10, rttol = 3, rt = c(min(MS1$RT), max(MS1$RT)),
  adducts = c("M+Na", "M+H"), clfrags = c("dg_M+H-H2O"),
  clrequired = c(F), ftype = c("BB"), chainfrags_sn1 = c("lysope_M+H-H2O",
  "mg_M+H-H2O"), chainfrags_sn2 = c("fa_M+H-H2O", "mg_M+H-H2O"),
  intrules = c("mg_sn1/mg_sn1", "lysope_sn1/lysope_sn1"), rates = c("3/1",
  "3/1"), intrequired = c(T), dbs = list(pedb = LipidMS::pedb, dgdb =
  LipidMS::dgdb, mgdb = LipidMS::mgdb, lysopedb = LipidMS::lysopedb, fadb =
  LipidMS::fadb, adductsTable = LipidMS::adductsTable))
}
\arguments{
\item{MS1}{data frame cointaining all peaks from the full MS function. It
must have three columns: m.z, RT (in seconds) and int (intensity).}

\item{MSMS1}{data frame cointaining all peaks from the low energy function.
It must have three columns: m.z, RT and int.}

\item{MSMS2}{data frame cointaining all peaks from the high energy
function if it is the case. It must have three columns: m.z, RT and int.
Optional.}

\item{ppm_precursor}{mass tolerance for precursor ions. By default, 5 ppm.}

\item{ppm_products}{mass tolerance for product ions. By default, 10 ppm.}

\item{rttol}{total rt window for coelution between precursor and product
ions. By default, 3 seconds.}

\item{rt}{rt range where the function will look for candidates. By default,
it will search within all RT range in MS1.}

\item{adducts}{expected adducts for PE in ESI+. Adducts allowed can
be modified in adductsTable (dbs argument).}

\item{clfrags}{vector containing the expected fragments for a given lipid
class. See \link{checkClass} for details.}

\item{clrequired}{logical vector indicating if each class fragment is
required or not. If any of them is required, at least one of them must be
present within the coeluting fragments. See \link{checkClass} for details.}

\item{ftype}{character vector indicating the type of fragments in clfrags.
It can be: "F" (fragment), "NL" (neutral loss) or "BB" (building block).
See \link{checkClass} for details.}

\item{chainfrags_sn1}{character vector containing the fragmentation rules for
the chain fragments in sn1 position. See \link{chainFrags} for details.}

\item{chainfrags_sn2}{character vector containing the fragmentation rules for
the chain fragments in sn2 position. See \link{chainFrags} for details. If
empty, it will be estimated based on the difference between precursors and
sn1 chains.}

\item{intrules}{character vector specifying the fragments to compare. See
\link{checkIntensityRules}.}

\item{rates}{character vector with the expected rates between fragments given
as a string (i.e. "3/1"). See \link{checkIntensityRules}.}

\item{intrequired}{logical vector indicating if any of the rules is required.
If not, at least one must be verified to confirm the structure.}

\item{dbs}{list of data bases required for annotation. By default, dbs
contains the required data frames based on the default fragmentation rules.
If these rules are modified, dbs may need to be changed. If data bases have
been customized using \link{createLipidDB}, they also have to be modified
here.}
}
\value{
List with PE annotations (results) and some additional information
(class fragments and chain fragments).
}
\description{
PE identification based on fragmentation patterns for LC-MS/MS
AIF data acquired in positive mode.
}
\details{
\code{idPEpos} function involves 5 steps. 1) FullMS-based
identification of candidate PE as M+H and M+Na. 2) Search of PE class
fragments: loss of head group (DG as M+H-H2O) coeluting with the precursor
ion. 3) Search of specific fragments that inform about chain composition at
sn1 (MG as M+H-H2O resulting from the loss of the FA chain at sn2 and
the head group or LPE as M+H-H2O resulting just from the loss of the FA chain)
and sn2 (FA or MG chain from sn2as M+H-H2O). 4) Look for possible chains
structure based on the combination of chain fragments. 5) Check intensity
rules to confirm chains position. LPE or MG from sn1 is at least 3 times more
intense than the ones from sn2.

Results data frame shows: ID, class of lipid, CDB (total number
of carbons and double bounds), FA composition (specific chains composition if
it has been confirmed), mz, RT (in seconds), I (intensity, which comes
directly from de input), Adducts, ppm (m.z error), confidenceLevel (Subclass,
FA level, where chains are known but not their positions, or FA position
level).
}
\note{
Isotopes should be removed before identification to avoid false
positives.
This function has been writen based on fragmentation patterns observed for
two different platforms (QTOF 6550 from Agilent and Sinapt G2-Si from Waters),
but it may need to be customized for other platforms or acquisition settings.
}
\examples{
idPEpos(MS1 = LipidMS::mix_pos_fullMS, MSMS1 = LipidMS::mix_pos_Ce20,
MSMS2 = LipidMS::mix_pos_Ce40)

idPEpos(MS1 = LipidMS::serum_pos_fullMS, MSMS1 = LipidMS::serum_pos_Ce20,
MSMS2 = LipidMS::serum_pos_Ce40)

}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
