% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idSphpos.R
\name{idSphpos}
\alias{idSphpos}
\title{Sphingoid bases (Sph) annotation for ESI-}
\usage{
idSphpos(MS1, MSMS1, MSMS2 = data.frame(), ppm_precursor = 5,
  ppm_products = 10, rttol = 3, rt = c(min(MS1$RT), max(MS1$RT)),
  adducts = c("M+H"), clfrags = c(), clrequired = c(), ftype = c(),
  chainfrags_sn1 = c("sph_M+H-H20", "sph_M+H-2H2O"), dbs = list(sphdb =
  LipidMS::sphdb, adductsTable = LipidMS::adductsTable))
}
\arguments{
\item{MS1}{data frame cointaining all peaks from the full MS function. It
must have three columns: m.z, RT (in seconds) and int.}

\item{MSMS1}{data frame cointaining all peaks from the first high energy
function. It must have three columns: m.z, RT (in seconds) and int.}

\item{MSMS2}{data frame cointaining all peaks from the second high energy
function if it is the case. It must have three columns: m.z, RT (in seconds)
and int. Optional.}

\item{ppm_precursor}{mass tolerance for precursor ions. By default, 5 ppm.}

\item{ppm_products}{mass tolerance for product ions. By default, 10 ppm.}

\item{rttol}{total rt window for coelution between precursors and product
ions. By default, 3 seconds.}

\item{rt}{rt window where the function will look for candidates. By default,
it will search within all RT range in MS1.}

\item{adducts}{expected adducts for Sph in ESI+. Adducts allowed can
be modified in adductsTable (dbs argument).}

\item{clfrags}{vector containing the expected fragments for a given lipid
class. See \link{checkClass} for details.}

\item{clrequired}{logical vector indicating if each class fragment is
required or not. If any of them is required, at least one of them must be
present within the coeluting fragments. See \link{checkClass} for details.}

\item{ftype}{character vector indicating the type of fragments in clfrags.
It can be: "F" (fragment), "NL" (neutral loss) or "BB" (building block).
See \link{checkClass} for details.}

\item{chainfrags_sn1}{character vector containing the fragmentation rules for
the chain fragments. See \link{chainFrags} for details.}

\item{dbs}{list of data bases required for the annotation. By default, dbs
contains the required data frames based on the default fragmentation rules.
If these rules are modified, dbs may need to be changed. If data bases have
been customized using \link{createLipidDB}, they also have to be uploaded here.}
}
\value{
list with Sph annotations (results) and some additional information
(class fragments and chain fragments).
}
\description{
Sph identification based on fragmentation patterns for LC-MS/MS
AIF data acquired in positive mode.
}
\details{
\code{idSphpos} function involves 3 steps. 1) FullMS-based
identification of candidate Sph as M+H. 2) Search of Sph class fragments in
case some is defined. 3) Search of specific fragments that confirm chain
composition (neutral loss of 1 or 2 H2O molecules).

Results data frame shows: ID, class of lipid, CDB (total number
of carbons and double bounds), FA composition (specific chains composition if
it has been confirmed), mz, RT (in seconds), I (intensity, which comes
directly from de input), Adducts, ppm (m.z error), confidenceLevel (in this
case, as Sph only have one chain, only Subclass and FA level are possible).
}
\note{
Isotopes should be removed before identification to avoid false
positives.
This function has been writen based on fragmentation patterns observed for
a Q-TOF 6550 from Agilent.
}
\examples{
idSphpos(MS1 = LipidMS::mix_pos_fullMS, MSMS1 = LipidMS::mix_pos_Ce20,
MSMS2 = LipidMS::mix_pos_Ce40)

idSphpos(MS1 = LipidMS::serum_pos_fullMS, MSMS1 = LipidMS::serum_pos_Ce20,
MSMS2 = LipidMS::serum_pos_Ce40)

}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
