% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalProcessing.R
\name{indexrtpart}
\alias{indexrtpart}
\title{Index partitions or clusters assigned during alignment.}
\usage{
indexrtpart(peaks, part, minsamples)
}
\arguments{
\item{peaks}{peaks obtained using \link{getallpeaks}.}

\item{part}{numeric vector with the partition/cluster assigned to each peak.}

\item{minsamples}{minimum number of samples represented in each partition.}
}
\value{
list with two elements: index and vector with the assigned partitions.
}
\description{
Index partitions or clusters assigned during alignment.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
