% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{joinAdducts}
\alias{joinAdducts}
\title{Join information about precursor candidates found using different adducts}
\usage{
joinAdducts(df1, df2, rttol, main, other)
}
\arguments{
\item{df1}{data frame containing identification results using the main adduct}

\item{rttol}{retention time tolerance in seconds}

\item{main}{character value indicating the preferred adduct}

\item{other}{character value indicating the ceecondary adduct}

\item{df}{data frame containing identification results using a secondary
adduct}
}
\value{
Subset of the original data frame without adducts
}
\description{
This function joins tables of precursor candidates found using different
adducts.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
