% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{joinAnnotationResults}
\alias{joinAnnotationResults}
\title{Summarize annotation results from an msbatch into the features table}
\usage{
joinAnnotationResults(msbatch, simplifyAnnotations = TRUE)
}
\arguments{
\item{msbatch}{msbatch}

\item{simplifyAnnotations}{logical. If TRUE, only the most frequent id will be 
kept (recommended when only pool samples have been acquired in DIA or DDA). If 
FALSE, all annotations will be shown.}
}
\value{
msbatch
}
\description{
Summarize annotation results from an msbatch into the features table
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
