\name{lysopedb}
\alias{lysopedb}
\docType{data}
\title{
LPEs database
}
\description{
In silico generated database for common LPEs.
}
\usage{data("lysopedb")}
\format{
  Data frame with 30 observations and the following 3 variables.
  \describe{
    \item{\code{formula}}{character vector containing molecular formulas.}
    \item{\code{total}}{character vector indicating the total number of carbons
    and double bounds of the chains.}
    \item{\code{Mass}}{numeric vector with the neutral masses.}
  }
}
\keyword{datasets}
