% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassCer}
\alias{MassCer}
\title{Calculate formula and mass of ceramides}
\usage{
MassCer(cer)
}
\arguments{
\item{cer}{character value indicating total number of carbons and double
bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of ceramides
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
