% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassCL}
\alias{MassCL}
\title{Calculate formula and mass of CL}
\usage{
MassCL(CL)
}
\arguments{
\item{CL}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of CL
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
