% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassDG}
\alias{MassDG}
\title{Calculate formula and mass of DG}
\usage{
MassDG(DG)
}
\arguments{
\item{DG}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of DG
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
