% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassFAHFA}
\alias{MassFAHFA}
\title{Calculate formula and mass of FAHFA}
\usage{
MassFAHFA(FAHFA)
}
\arguments{
\item{FAHFA}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of FAHFA
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
