% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassGlcCer}
\alias{MassGlcCer}
\title{Calculate formula and mass of glucoceramides}
\usage{
MassGlcCer(glccer)
}
\arguments{
\item{glccer}{character value indicating total number of carbons and double
bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of glucoceramides
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
