% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassHFA}
\alias{MassHFA}
\title{Calculate formula and mass of hydroxi fatty acids}
\usage{
MassHFA(HFA)
}
\arguments{
\item{HFA}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of hydroxi fatty acids
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
