% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassLysoPA}
\alias{MassLysoPA}
\title{Calculate formula and mass of LPA}
\usage{
MassLysoPA(LPA)
}
\arguments{
\item{LPA}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of LPA
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
