% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassLysoPC}
\alias{MassLysoPC}
\title{Calculate formula and mass of LysoPC}
\usage{
MassLysoPC(LPC)
}
\arguments{
\item{LPC}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of LysoPC
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
