% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassLysoPE}
\alias{MassLysoPE}
\title{Calculate formula and mass of LPE}
\usage{
MassLysoPE(LPE)
}
\arguments{
\item{LPE}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of LPE
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
