% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassLysoPS}
\alias{MassLysoPS}
\title{Calculate formula and mass of LysoPS}
\usage{
MassLysoPS(LPS)
}
\arguments{
\item{LPS}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of LysoPS
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
