% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassMG}
\alias{MassMG}
\title{Calculate formula and mass of MG}
\usage{
MassMG(MG)
}
\arguments{
\item{MG}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of MG
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
