% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassPC}
\alias{MassPC}
\title{Calculate formula and mass of PC}
\usage{
MassPC(PC)
}
\arguments{
\item{PC}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of PC
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
