% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassPEo}
\alias{MassPEo}
\title{Calculate formula and mass of plasmanyl PE}
\usage{
MassPEo(PEo)
}
\arguments{
\item{PEo}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of plasmanyl PE
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
