% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassPG}
\alias{MassPG}
\title{Calculate formula and mass of PG}
\usage{
MassPG(PG)
}
\arguments{
\item{PG}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of PG
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
