% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassSph}
\alias{MassSph}
\title{Calculate formula and mass of sphingoid bases}
\usage{
MassSph(Sph)
}
\arguments{
\item{Sph}{character value indicating total number of carbons and double
bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of sphingoid bases
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
