% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassSphP}
\alias{MassSphP}
\title{Calculate formula and mass of sphingoid phosphate bases}
\usage{
MassSphP(SphP)
}
\arguments{
\item{SphP}{character value indicating total number of carbons and double
bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of sphingoid phosphate bases
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
