\name{mix_neg_Ce20}
\alias{mix_neg_Ce20}
\docType{data}
\title{
Peak list example of MSMS1 obtained from a mix of standards.
}
\description{
Data frame containing the MSMS1 peak list of a mix of standards.
}
\usage{data("mix_neg_Ce20")}
\format{
  A data frame with 17315 observations on the following 3 variables.
  \describe{
    \item{\code{m.z}}{a numeric vector}
    \item{\code{RT}}{a numeric vector}
    \item{\code{int}}{a numeric vector}
  }
}
\details{
Q-TOF 6550 from Agilent.
MS/MS DIA acquisition.
ESI -.
Collision energy: 20.
Preprocessed with enviPick(peak-picking) and xcms/CAMERA (isotopes removal).
}
\keyword{datasets}
