\name{mix_neg_Ce40}
\alias{mix_neg_Ce40}
\docType{data}
\title{
Peak list example of MSMS2 obtained from a mix of standards.
}
\description{
Data frame containing the MSMS2 peak list of a mix of standards.
}
\usage{data("mix_neg_Ce40")}
\format{
  A data frame with 22457 observations on the following 3 variables.
  \describe{
    \item{\code{m.z}}{a numeric vector}
    \item{\code{RT}}{a numeric vector}
    \item{\code{int}}{a numeric vector}
  }
}
\details{
Q-TOF 6550 from Agilent.
MS/MS DIA acquisition.
ESI -.
Collision energy: 40.
Preprocessed with enviPick(peak-picking) and xcms/CAMERA (isotopes removal).
}
\keyword{datasets}
