\name{nlsphdb}
\alias{nlsphdb}
\docType{data}
\title{
Neutral losses db for sphingoid bases. It is employed by idCerneg function.
}
\description{
In silico generated database for neutral losses of sphingoid bases in ESI-.
}
\usage{data("nlsphdb")}
\format{
  Data frame with 4 observations and the following 3 variables.
  \describe{
    \item{\code{formula}}{character vector containing molecular formulas.}
    \item{\code{total}}{character vector indicating the total number of carbons
    and double bounds of the chains.}
    \item{\code{Mass}}{numeric vector with the neutral masses.}
  }
}
\keyword{datasets}
