\name{padb}
\alias{padb}
\docType{data}
\title{
PAs database
}
\description{
In silico generated database for common PAs.
}
\usage{data("padb")}
\format{
  Data frame with 147 observations and the following 3 variables.
  \describe{
    \item{\code{formula}}{character vector containing molecular formulas.}
    \item{\code{total}}{character vector indicating the total number of carbons
    and double bounds of the chains.}
    \item{\code{Mass}}{numeric vector with the neutral masses.}
  }
}
\keyword{datasets}
