% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalProcessing.R
\name{partitioning}
\alias{partitioning}
\title{agglomarative partitioning for LC-HRMS data based on enviPick algorithm}
\usage{
partitioning(msobject, dmzagglom, drtagglom, minpeak, mslevel, cE)
}
\arguments{
\item{msobject}{msobject generated by \link{readMSfile}}

\item{dmzagglom}{mz tolerance for partitions}

\item{drtagglom}{RT window for partitions}

\item{minpeak}{minimum number of measures to define a peak}

\item{mslevel}{MS level information to access msobject}

\item{cE}{collision energy information to access msobject}
}
\value{
msobject
}
\description{
agglomarative partitioning for LC-HRMS data based on enviPick algorithm
}
\references{
Peak-picking algorithm has been imported from enviPick R-package:
https://cran.r-project.org/web/packages/enviPick/index.html
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
