% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalFunctions.R
\name{ploteicmsbatch}
\alias{ploteicmsbatch}
\title{EIC for all samples in a msbatch}
\usage{
ploteicmsbatch(msbatch, mz, ppm, rt, colorbygroup = TRUE, verbose = TRUE)
}
\arguments{
\item{msbatch}{msbatch}

\item{mz}{mz of interest}

\item{ppm}{mass tolerance in ppm}

\item{rt}{numeric vector with the RT range to be plotted}

\item{colorbygroup}{logical. If TRUE, samples will be coloured based on their
sample group (from metadata).}

\item{verbose}{print information messages.}
}
\value{
plot
}
\description{
EIC for all samples in a msbatch
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
