% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalFunctions.R
\name{plotticmsbatch}
\alias{plotticmsbatch}
\title{TIC for all samples in a msbatch}
\usage{
plotticmsbatch(msbatch, rt, colorbygroup = TRUE)
}
\arguments{
\item{msbatch}{msbatch}

\item{rt}{numeric vector with the RT range to be plotted}

\item{colorbygroup}{logical. If TRUE, samples will be coloured based on their
sample group (from metadata).}
}
\value{
plot
}
\description{
TIC for all samples in a msbatch
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
