% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{readFileLipidClasses}
\alias{readFileLipidClasses}
\title{Allows reading a CSV that contains lipid classes and adducts to perform
annotations and retention time predictions.}
\usage{
readFileLipidClasses(file = NULL, sep = ";", sep_adducts = ",")
}
\arguments{
\item{file}{csv with classes to predict}

\item{sep}{Character. Field separator used when reading the CSV file.}

\item{sep_adducts}{Character. Separator used to split multiple adducts within the "Adducts" column of the file.}
}
\value{
lipidClasses, a list with Classes and Adducts
}
\description{
Allows reading a CSV that contains lipid classes and adducts to perform
annotations and retention time predictions.
}
\author{
Celia Regal Marques <celia_regal@iislafe.es>
}
\keyword{internal}
