% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalProcessing.R
\name{readMSfile}
\alias{readMSfile}
\title{Read .mzXML files and initiate msobject}
\usage{
readMSfile(file)
}
\arguments{
\item{file}{file path for a .mzXML file}
}
\value{
msobject
}
\description{
Read .mzXML files and initiate msobject
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
