% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{removeAdducts}
\alias{removeAdducts}
\title{remove wrong assigned adducts}
\usage{
removeAdducts(df, mdiff, ppm, rttol)
}
\arguments{
\item{df}{data frame with the input results}

\item{mdiff}{mass difference expected between adducts}

\item{ppm}{mass tolerance}

\item{rttol}{rt tolerance}
}
\value{
Data frame
}
\description{
remove wrong assigned adducts.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
