% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{removeDGs}
\alias{removeDGs}
\title{remove fragments whose relative intensity to the precursor is less than 10
times smaller}
\usage{
removeDGs(sn, candidates)
}
\arguments{
\item{sn}{fragments}

\item{candidates}{precursors}
}
\value{
Subsetted data frame
}
\description{
remove fragments whose relative intensity to the precursor is less than 10
times smaller. It is used by idTGpos to reduce false positive annotations.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
