% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalFunctions.R
\name{rtdevplot}
\alias{rtdevplot}
\title{Plot retention time deviation}
\usage{
rtdevplot(msbatch, colorbygroup = TRUE)
}
\arguments{
\item{msbatch}{aligned msbatch.}

\item{colorbygroup}{logical. If TRUE, samples will be coloured based on their
sample group (from metadata).}
}
\value{
plot
}
\description{
Plot retention time deviation of an aligned msbatch
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
