% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{select}
\alias{select}
\title{Check matches between chains composition and precursor structures}
\usage{
select(chains, parent, n)
}
\arguments{
\item{chains}{character value containing chains structure separated by a
white space.}

\item{parent}{precursor ion structure}

\item{n}{number of chains}
}
\value{
Logical value
}
\description{
This function checks if the sum up of the chains structure match the
precursor structure. It is used by \link{combineChains}.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
