% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalProcessing.R
\name{annotateIsotopes}
\alias{annotateIsotopes}
\title{Annotate isotopes}
\usage{
annotateIsotopes(peaklist, rawScans, dmzIso, drtIso)
}
\arguments{
\item{peaklist}{extracted peaks. Data.frame with 4 columns (m.z, RT, int
and peakID).}

\item{rawScans}{raw scan data. Data.frame with 5 columns (m.z, RT, int,
peakID and Scan).}

\item{dmzIso}{mass tolerance in ppm.}

\item{drtIso}{rt windows with the same units used in peaklist.}
}
\value{
peaklist with 6 columns (m.z, RT, int, peakID, isotope and group).
}
\description{
Annotate isotopes based on mass differences, retention time and peak
correlation if required.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@iislafe.es>
}
\keyword{internal}
