% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalFunctions.R
\name{assignDB}
\alias{assignDB}
\title{Load LipidMS default data bases}
\usage{
assignDB()
}
\value{
list of data frames
}
\description{
load all LipidMS default data bases required to run identification functions.
}
\examples{
\dontrun{
dbs <- assignDB()
}

}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
