% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{chains}
\alias{chains}
\title{extract chains composition from a lipid name}
\usage{
chains(id)
}
\arguments{
\item{id}{lipid name}
}
\value{
vector with lipid class, FA position known (TRUE) or unknown (FALSE) 
and FA chains.
}
\description{
extract chains composition from a lipid name
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
