% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalProcessing.R
\name{clustdist}
\alias{clustdist}
\title{Calculate max distance between clusters.}
\usage{
clustdist(mins, maxs)
}
\arguments{
\item{mins}{lower bound for each cluster.}

\item{maxs}{higher bound for each cluster.}
}
\value{
numeric vector with the assigned clusters
}
\description{
Calculate max distance between clusters.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@iislafe.es>
}
\keyword{internal}
