% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coelutingFrags.R
\name{coelutingFrags}
\alias{coelutingFrags}
\title{Coeluting fragments extraction}
\usage{
coelutingFrags(precursors, products, rttol, rawData = data.frame(),
  coelCutoff = 0)
}
\arguments{
\item{precursors}{candidates data frame. Output of \link{findCandidates}.}

\item{products}{peaklist for MS2 function (MSMS).}

\item{rttol}{rt window in seconds.}

\item{rawData}{raw scans data. Output of \link{dataProcessing} function
(MSMS$rawData).}

\item{coelCutoff}{coelution score threshold between parent and fragment ions.
Only applied if rawData info is supplied.}
}
\value{
List of data frames with the coeluting fragments for each candidate.
}
\description{
Given a RT and a list of peaks, this function subsets all coeluting fragments
within a rt windows. It is used by identification functions to extract
coeluting fragments from high energy functions for candidate precursor ions.
}
\examples{
\donttest{
library(LipidMSdata)
dbs <- assignDB()

candidates <- findCandidates(MS1 = MS1_neg$peaklist,
db = dbs$pgdb, ppm = 10, rt = c(0, 2000), adducts = c("M-H"),
rttol = 10, rawData = MS1_neg$rawScans, coelCutoff = 0.8)

MSMS <- rbind(MSMS1_neg$peaklist, MSMS2_neg$peaklist)
rawData <- rbind(MS1_neg$rawScans, MSMS1_neg$rawScans, MSMS2_neg$rawScans)
coelfrags <- coelutingFrags(candidates$RT, MSMS, rttol = 10, rawData = rawData,
coelCutoff = 0.8)
}

}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
