% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{coelutionScore}
\alias{coelutionScore}
\title{calculate coelution score between two peaks}
\usage{
coelutionScore(peak1, peak2, rawData)
}
\arguments{
\item{peak1}{character vector specifying the peakID of the first peak.}

\item{peak2}{character vector specifying the peakID of the second peak.}

\item{rawData}{data frame with raw data for each scan. it need to have at
least 5 columns: mz, RT, int, Scan (ordinal number for a given MS function)
and peakID (peakID to which it has been assigned).

#' @keywords internal}
}
\description{
Calculate coelution score between two peaks.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
