% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{dbOneChain}
\alias{dbOneChain}
\title{Creation of a database for Carnitines, CE, FA, HFA, LPL, MG, sphingoid
bases and sphingoid bases phosphate.}
\usage{
dbOneChain(chains, lipid)
}
\arguments{
\item{chains}{character vector indicating the FAs to be employed}

\item{lipid}{character value indication the class of lipid.}
}
\value{
data frame containing formula, mass and total number of carbons and
insaturations
}
\description{
Creation of a database for Carnitines, CE, FA, HFA, LPL, MG, sphingoid
bases and sphingoid bases phosphate.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
