% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{dbPlasmalogens}
\alias{dbPlasmalogens}
\title{Creation of a database for PCo, PCp, PEo and PEp}
\usage{
dbPlasmalogens(chains, chains2, lipid)
}
\arguments{
\item{chains}{character vector indicating the FAs to be employed for sn1}

\item{chains2}{character vector indicating the FAs to be employed for sn2}

\item{lipid}{character value indication the class of lipid.}
}
\value{
data frame containing formula, mass and total number of carbons and
insaturations.
}
\description{
Creation of a database for PCo, PCp, PEo and PEp
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
