% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{dbSphingolipids}
\alias{dbSphingolipids}
\title{Creation of a database for Cer, CerP, GlcCer and SM}
\usage{
dbSphingolipids(chains, chains2, lipid)
}
\arguments{
\item{chains}{character vector indicating the FAs to be employed}

\item{chains2}{character vector indicating the sphingoid bases to be employed}

\item{lipid}{character value indication the class of lipid.}
}
\value{
data frame containing formula, mass and total number of carbons and
insaturations.
}
\description{
Creation of a database for Cer, CerP, GlcCer and SM
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
