% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{fillpeaksmsbatch}
\alias{fillpeaksmsbatch}
\title{Fill peaks from a grouped msbatch}
\usage{
fillpeaksmsbatch(msbatch)
}
\arguments{
\item{msbatch}{msbatch obtained from the \link{groupmsbatch} function.}
}
\value{
msbatch
}
\description{
Use grouping results to target all peaks from the msbatch in each sample and 
refill intensities at the features table.
}
\details{
Once grouping has been performed, areas are extracted again for each 
peak and sample based on the peak parameters defined for each feature (mz and 
tolerance and initial and final RT).
}
\examples{
\dontrun{
msbatch <- fillpeaksmsbatch(msbatch)
}

}
\author{
M Isabel Alcoriza-Balaguer <maialba@iislafe.es>
}
