% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{filtermsms}
\alias{filtermsms}
\title{Presence or absence of an mz value withing a vector of mz values}
\usage{
filtermsms(fragments, frag, ppm)
}
\arguments{
\item{fragments}{vector of mz values}

\item{frag}{mz to be matched}

\item{ppm}{mass tolerance}
}
\value{
Logical value
}
\description{
This function indicates the presence or absence of a fragment within a set
of mz values with certain tolerance. It is used by identification functions
to look for the generic fragments of each class of lipid.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
