% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{filtrateAdducts}
\alias{filtrateAdducts}
\title{Remove low adduct supported candidates to avoid false positives.}
\usage{
filtrateAdducts(df)
}
\arguments{
\item{df}{data frame containing candidates}
}
\value{
Subset of the original data frame
}
\description{
In case some feature has been annotated to different candidate species,
this function removes the one with less adducts assigned.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
