% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{findPrecursor}
\alias{findPrecursor}
\title{Find candidate precursor from fullMS function}
\usage{
findPrecursor(MS1, db, ppm, massdif, rt, n = 1, charge = 1)
}
\arguments{
\item{MS1}{data frame containing m/z, RT and intensity of ions from the first
function}

\item{db}{database to be searched for matches}

\item{massdif}{mass difference between neutral mass and the adduct expected}

\item{rt}{rt window}

\item{n}{numeric value indicating whether to look for monomers (1), dimers
(2), etc.}

\item{charge}{numeric value indicating the expected charge of the ions}
}
\value{
Subset of the original data frame without adducts
}
\description{
This function is employed by all identification function in this package to
find possible precursors in the fullMS function.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
