% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{getFormula}
\alias{getFormula}
\title{Get formula and neutral mass for annotated compounds}
\usage{
getFormula(df, dbs)
}
\arguments{
\item{df}{data frame with the input results}

\item{dbs}{list of data bases required for annotation. By default, dbs
contains the required data frames based on the default fragmentation rules.
If these rules are modified, dbs may need to be supplied. See \link{createLipidDB}
and \link{assignDB}.}
}
\value{
Data frame
}
\description{
Get formula and neutral mass for annotated compounds.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
