% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{getFormula}
\alias{getFormula}
\title{Get formula and neutral mass for annotated compounds}
\usage{
getFormula(df)
}
\arguments{
\item{df}{data frame with the input results}
}
\value{
Data frame
}
\description{
Get formula and neutral mass for annotated compounds.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
