% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalFunctions.R
\name{getInclusionList}
\alias{getInclusionList}
\title{Obtain an inclusion list from the annotation results}
\usage{
getInclusionList(df, dbs)
}
\arguments{
\item{df}{data frame. Output of identification functions (results table from
an msobject or feature table from an msbatch).}

\item{dbs}{list of data bases required for annotation. By default, dbs
contains the required data frames based on the default fragmentation rules.
If these rules are modified, dbs may need to be supplied. See \link{createLipidDB}
and \link{assignDB}.}
}
\value{
Data frame with 6 columns: formula, RT, neutral mass, m/z, adduct
and the LipidMSid.
}
\description{
Obtain an inclusion list for the identified lipids.
}
\examples{
\dontrun{
devtools::install_github("maialba3/LipidMSdata2")

library(LipidMS)
msobject <- idPOS(LipidMSdata2::msobjectDIApos)
getInclusionList(msobject)
}

}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
