% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idFunctionsNeg.R
\name{idAcylCerneg}
\alias{idAcylCerneg}
\title{Acylceramides (AcylCer) annotation for ESI-}
\usage{
idAcylCerneg(
  msobject,
  ppm_precursor = 5,
  ppm_products = 10,
  rttol = 3,
  rt,
  adducts = c("M-H", "M+CH3COO"),
  clfrags = c(),
  clrequired = c(),
  ftype = c(),
  chainfrags_sn1 = c("cbdiff-cer_M-H"),
  chainfrags_sn2 = c("sph_Mn-62.06001", "sph_M-H-H2O"),
  chainfrags_sn3 = c("fa_Mn-1.9918", "fa_Mn-19.0179"),
  intrules = c("cbdiff-cer_sn1/sph_sn2", "sph_sn2/fa_sn3"),
  rates = c("5/1", "2/1"),
  intrequired = c(T, T),
  coelCutoff = 0.8,
  dbs
)
}
\arguments{
\item{msobject}{an msobject returned by \link{dataProcessing}.}

\item{ppm_precursor}{mass tolerance for precursor ions. By default, 5 ppm.}

\item{ppm_products}{mass tolerance for product ions. By default, 10 ppm.}

\item{rttol}{total rt window for coelution between precursor and product
ions. By default, 3 seconds.}

\item{rt}{rt range where the function will look for candidates. By default,
it will search within all RT range in MS1.}

\item{adducts}{expected adducts for AcylCer in ESI-. Adducts allowed can
be modified in adductsTable (dbs argument).}

\item{clfrags}{vector containing the expected fragments for a given lipid
class. See \link{checkClass} for details.}

\item{clrequired}{logical vector indicating if each class fragment is
required or not. If any of them is required, at least one of them must be
present within the coeluting fragments. See \link{checkClass} for details.}

\item{ftype}{character vector indicating the type of fragments in clfrags.
It can be: "F" (fragment), "NL" (neutral loss) or "BB" (building block).
See \link{checkClass} for details.}

\item{chainfrags_sn1}{character vector containing the fragmentation rules for
the sphingoid base. See \link{chainFrags} for details.}

\item{chainfrags_sn2}{character vector containing the fragmentation rules for
the chain fragments in sn2 position. See \link{chainFrags} for details. If
empty, it will be estimated based on the difference between precursors and
sn1 chains.}

\item{chainfrags_sn3}{character vector containing the fragmentation rules for
the acyl chain. See \link{chainFrags} for details.}

\item{intrules}{character vector specifying the fragments to compare. See
\link{checkIntensityRules}.}

\item{rates}{character vector with the expected ratesbetween fragments given
as a string (e.g. "3/1"). See \link{checkIntensityRules}.}

\item{intrequired}{logical vector indicating if any of the rules is required.
If not, at least one must be verified to confirm the structure.}

\item{coelCutoff}{coelution score threshold between parent and fragment ions.
Only applied if rawData info is supplied. By default, 0.8.}

\item{dbs}{list of data bases required for annotation. By default, dbs
contains the required data frames based on the default fragmentation rules.
If these rules are modified, dbs may need to be supplied. See \link{createLipidDB}
and \link{assignDB}.}
}
\value{
annotated msobject (list with several elements). The results element
is a data frame that shows: ID, lipid class, CDB (total number of carbons
and double bounds), FA composition (specific chains composition if it has
been confirmed), mz, RT (in seconds), I (intensity), Adducts, ppm (mz error),
confidenceLevel (Subclass, FA level, where chains are known but not their
positions, or FA position level), peakID, and Score (parent-fragment coelution 
score mean in DIA data or relative sum intensity in DDA of all fragments used 
for the identification).
}
\description{
AcylCer identification based on fragmentation patterns for LC-MS/MS DIA or DDA
data acquired in negative mode.
}
\details{
\code{idAcylCerneg} function involves 5 steps. 1) FullMS-based
identification of candidate AcylCer as M-H and M+CH3COO. 2) Search of AcylCer 
class fragments: no class fragments by default. 3) Search of specific fragments 
that inform about the acyl chain (Cer as M-H), the sphingoid base (neutral 
loss of 62.0600 of the Sph) and the FA chain (FA as M-H and M-H2O but with a N 
instead of an O, what results in a mass differences of 1.9918 and 19.0179 
respectively). 4) Look for possible chains structure based on the combination 
of chain fragments. 5) Check intensity rules to confirm chains position. In 
this case, the fragment coming from the loss of the acyl chain must be at least 
5 times more intense the fragment from the sphingoid base and this one, two 
times more intense than the FA chain from sn3.

Results data frame shows: ID, lipid class, CDB (total number
of carbons and double bounds), FA composition (specific chains composition if
it has been confirmed), mz, RT (in seconds), I (intensity, which comes
directly from de input), Adducts, ppm (mz error), confidenceLevel (Subclass,
FA level, where chains are known but not their positions, or FA position
level) and Score (parent-fragment coelution score mean in DIA data or relative 
sum intensity in DDA of all fragments used for the identification).
}
\note{
This function has been writen based on fragmentation patterns
observed for three different platforms (QTOF 6550 from Agilent, Sinapt G2-Si
from Waters and Q-exactive from Thermo), but it may need to be customized for
other platforms or acquisition settings.
}
\examples{
\dontrun{
msobject <- idAcylCerneg(msobject)
}

}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
