% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idFunctionsNeg.R
\name{idPEneg}
\alias{idPEneg}
\title{Phosphoethanolamines (PE) annotation for ESI-}
\usage{
idPEneg(
  msobject,
  ppm_precursor = 5,
  ppm_products = 10,
  rttol = 5,
  rt,
  adducts = c("M-H"),
  clfrags = c(140.0118, 196.038, 214.048, "pe_M-CH3"),
  clrequired = c(F, F, F, "excluding"),
  ftype = c("F", "F", "F", "BB"),
  chainfrags_sn1 = c("lysope_M-H"),
  chainfrags_sn2 = c("lysope_M-H", "fa_M-H"),
  intrules = c("lysope_sn1/lysope_sn2"),
  rates = c("3/1"),
  intrequired = c(T),
  coelCutoff = 0.8,
  dbs
)
}
\arguments{
\item{msobject}{an msobject returned by \link{dataProcessing}.}

\item{ppm_precursor}{mass tolerance for precursor ions. By default, 5 ppm.}

\item{ppm_products}{mass tolerance for product ions. By default, 10 ppm.}

\item{rttol}{total rt window for coelution between precursor and product
ions. By default, 3 seconds.}

\item{rt}{rt range where the function will look for candidates. By default,
it will search within all RT range in MS1.}

\item{adducts}{expected adducts for PE in ESI-. Adducts allowed can
be modified in adductsTable (dbs argument).}

\item{clfrags}{vector containing the expected fragments for a given lipid
class. See \link{checkClass} for details.}

\item{clrequired}{logical vector indicating if each class fragment is
required or not. If any of them is required, at least one of them must be
present within the coeluting fragments. See \link{checkClass} for details.}

\item{ftype}{character vector indicating the type of fragments in clfrags.
It can be: "F" (fragment), "NL" (neutral loss) or "BB" (building block).
See \link{checkClass} for details.}

\item{chainfrags_sn1}{character vector containing the fragmentation rules for
the chain fragments in sn1 position. See \link{chainFrags} for details.}

\item{chainfrags_sn2}{character vector containing the fragmentation rules for
the chain fragments in sn2 position. See \link{chainFrags} for details. If
empty, it will be estimated based on the difference between precursors and
sn1 chains.}

\item{intrules}{character vector specifying the fragments to compare. See
\link{checkIntensityRules}.}

\item{rates}{character vector with the expected rates between fragments given
as a string (e.g. "3/1"). See \link{checkIntensityRules}.}

\item{intrequired}{logical vector indicating if any of the rules is required.
If not, at least one must be verified to confirm the structure.}

\item{coelCutoff}{coelution score threshold between parent and fragment ions.
Only applied if rawData info is supplied. By default, 0.8.}

\item{dbs}{list of data bases required for annotation. By default, dbs
contains the required data frames based on the default fragmentation rules.
If these rules are modified, dbs may need to be supplied. See \link{createLipidDB}
and \link{assignDB}.}
}
\value{
annotated msobject (list with several elements). The results element
is a data frame that shows: ID, class of lipid, CDB (total number of carbons
and double bounds), FA composition (specific chains composition if it has
been confirmed), m.z, RT (in seconds), I (intensity), Adducts, ppm (m.z error),
confidenceLevel (Subclass, FA level, where chains are known but not their
positions, or FA position level), peakID, and PFCS (parent-fragment coelution
score mean of all fragments used for the identification); and the
annotatedPeaklist element shows the original MS1 peaklist with the annotations
on it.
}
\description{
PE identification based on fragmentation patterns for LC-MS/MS DIA or DDA
data acquired in negative mode.
}
\details{
\code{idPEneg} function involves 5 steps. 1) FullMS-based
identification of candidate PE as M-H. 2) Search of PE class fragments:
140.0115, 196.038, 214.048 ion coeluting with the precursor ion. If a loss of
CH3 group is found coeluting with any candidate, this will be excluded as it
is a characteristic fragment of PC. 3) Search of specific fragments that
inform about chain composition in sn1 (lysoPE as M-H resulting from the loss
of the FA chain at sn2) and sn2 (lysoPE as M-H resulting from the loss
of the FA chain at sn1 or FA chain as M-H). 4) Look for possible
chains structure based on the combination of chain fragments. 5) Check
intensity rules to confirm chains position. In this case, lysoPE from sn1 is
at least 3 times more intense than lysoPE from sn2.

Results data frame shows: ID, class of lipid, CDB (total number
of carbons and double bounds), FA composition (specific chains composition if
it has been confirmed), mz, RT (in seconds), I (intensity, which comes
directly from de input), Adducts, ppm (m.z error), confidenceLevel (Subclass,
FA level, where chains are known but not their positions, or FA position
level) and PFCS (parent-fragment coelution score mean of all fragments used
for the identification).
}
\note{
This function has been writen based on fragmentation patterns
observed for three different platforms (QTOF 6550 from Agilent, Sinapt G2-Si
from Waters and Q-exactive from Thermo), but it may need to be customized for
other platforms or acquisition settings.
}
\examples{
\dontrun{
devtools::install_github("maialba3/LipidMSdata2")

library(LipidMS)
msobject <- idPEneg(LipidMSdata2::msobjectDIAneg)
}

}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
