% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{joinfrags}
\alias{joinfrags}
\title{Join fragments information when several peaks of the same fragment are
coeluting with a unique candidate}
\usage{
joinfrags(df)
}
\arguments{
\item{df}{data frame containing coeluting fragments}
}
\value{
Data frame
}
\description{
Function employed by \link{frags}.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
