% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMSapp.R
\name{LipidMSapp}
\alias{LipidMSapp}
\title{LipidMS shiny app}
\usage{
LipidMSapp()
}
\description{
Interactive UI for LipidMS
}
\examples{
\dontrun{
# example data files can be download from github.com/maialba3/LipidMSv2.0_exampleFiles

library(LipidMS)
LipidMSapp()
}

}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
