% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassAcylCer}
\alias{MassAcylCer}
\title{Calculate formula and mass of acylceramides}
\usage{
MassAcylCer(acylcer)
}
\arguments{
\item{acylcer}{character value indicating total number of carbons and double
bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of acylceramides
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
