% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassCarnitines}
\alias{MassCarnitines}
\title{Calculate formula and mass of carnitines}
\usage{
MassCarnitines(carnitine)
}
\arguments{
\item{carnitine}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of carnitines
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
