% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassCE}
\alias{MassCE}
\title{Calculate formula and mass of cholesterol esthers}
\usage{
MassCE(CE)
}
\arguments{
\item{CE}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of cholesterol esthers
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
