% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassCerP}
\alias{MassCerP}
\title{Calculate formula and mass of ceramides phosphate}
\usage{
MassCerP(cerP)
}
\arguments{
\item{cerP}{character value indicating total number of carbons and double
bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of ceramides phosphate
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
