% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassFA}
\alias{MassFA}
\title{Calculate formula and mass of fatty acids}
\usage{
MassFA(FA)
}
\arguments{
\item{FA}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of fatty acids
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
