% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassLysoPAo}
\alias{MassLysoPAo}
\title{Calculate formula and mass of LPAo}
\usage{
MassLysoPAo(LPAo)
}
\arguments{
\item{LPAo}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of LPAo
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
