% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassLysoPCo}
\alias{MassLysoPCo}
\title{Calculate formula and mass of LysoPCo}
\usage{
MassLysoPCo(LPCo)
}
\arguments{
\item{LPCo}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of LysoPCo
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
