% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassLysoPCp}
\alias{MassLysoPCp}
\title{Calculate formula and mass of LysoPCp}
\usage{
MassLysoPCp(LPCp)
}
\arguments{
\item{LPCp}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of LysoPCp
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
