% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassLysoPEo}
\alias{MassLysoPEo}
\title{Calculate formula and mass of LPEo}
\usage{
MassLysoPEo(LPEo)
}
\arguments{
\item{LPEo}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of LPEo
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
