% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassLysoPEp}
\alias{MassLysoPEp}
\title{Calculate formula and mass of LPEp}
\usage{
MassLysoPEp(LPEp)
}
\arguments{
\item{LPEp}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of LPEp
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
