% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassLysoPG}
\alias{MassLysoPG}
\title{Calculate formula and mass of LPG}
\usage{
MassLysoPG(LPG)
}
\arguments{
\item{LPG}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of LPG
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
