% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassLysoPI}
\alias{MassLysoPI}
\title{Calculate formula and mass of LPI}
\usage{
MassLysoPI(LPI)
}
\arguments{
\item{LPI}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of LPI
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
