% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassPA}
\alias{MassPA}
\title{Calculate formula and mass of PA}
\usage{
MassPA(PA)
}
\arguments{
\item{PA}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of PA
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
