% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassPCp}
\alias{MassPCp}
\title{Calculate formula and mass of PCp}
\usage{
MassPCp(PCp)
}
\arguments{
\item{PCp}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of PCp
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
