% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassPEp}
\alias{MassPEp}
\title{Calculate formula and mass of plasmenyl PE}
\usage{
MassPEp(PEp)
}
\arguments{
\item{PEp}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of plasmenyl PE
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
