% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassPIP2}
\alias{MassPIP2}
\title{Calculate formula and mass of PIP2}
\usage{
MassPIP2(PIP2)
}
\arguments{
\item{PIP2}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of PIP2
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
