% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassPIP3}
\alias{MassPIP3}
\title{Calculate formula and mass of PIP3}
\usage{
MassPIP3(PIP3)
}
\arguments{
\item{PIP3}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of PIP3
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
