% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassSM}
\alias{MassSM}
\title{Calculate formula and mass of sphingomyelines}
\usage{
MassSM(SM)
}
\arguments{
\item{SM}{character value indicating total number of carbons and double
bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of sphingomyelines
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
