% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internaldbs.R
\name{MassTG}
\alias{MassTG}
\title{Calculate formula and mass of TG}
\usage{
MassTG(TG)
}
\arguments{
\item{TG}{character value indicating total number of carbons and
double bounds}
}
\value{
vector containing formula and mass
}
\description{
Calculate formula and mass of TG
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
