% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{mzMatch}
\alias{mzMatch}
\title{mz match withing a vector of mz values}
\usage{
mzMatch(mz, mzvector, ppm)
}
\arguments{
\item{mz}{mz value to be matched}

\item{mzvector}{vector of mz values}

\item{ppm}{mass error tolerance}
}
\value{
Numeric vector indicating the index of matched mz values and ppms for
each one of those matches (match1, ppm1, match2, ppm2, etc.)
}
\description{
This function searches marches between a given mz and a vector of mz values
with certain mass  tolerance and returns the index of the matched values. It
is used by identification functions to find candidates of each class of lipid
based on full MS information.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
