% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalProcessing.R
\name{peakdetection}
\alias{peakdetection}
\title{peak-pick based on previous EIC clusters generated by \link{clustering}}
\usage{
peakdetection(
  msobject,
  minpeak,
  drtminpeak,
  drtmaxpeak,
  drtgap,
  recurs,
  weight,
  ended,
  sb,
  sn,
  minint,
  mslevel,
  cE
)
}
\arguments{
\item{msobject}{msobject generated by \link{clustering}}

\item{minpeak}{minimum number of measures to define a peak}

\item{drtminpeak}{minimum RT length of a peak}

\item{drtmaxpeak}{maximum RT length of a peak}

\item{drtgap}{maximum RT gap to be filled}

\item{recurs}{maximum number of peaks for a EIC}

\item{weight}{weight for assigning measurements to a peak}

\item{ended}{number of failures allowed when detecting peaks}

\item{sb}{signal-to-base ration}

\item{sn}{signal-to-noise ratio}

\item{minint}{minimum intensity}

\item{mslevel}{info to access msobject}

\item{cE}{info to access msobject}
}
\value{
msobject
}
\description{
peak-pick based on previous EIC clusters generated by \link{clustering}
}
\references{
Peak-picking algorithm has been imported from enviPick R-package:
https://cran.r-project.org/web/packages/enviPick/index.html
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
