% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalProcessing.R
\name{removeduplicatedpeaks}
\alias{removeduplicatedpeaks}
\title{Remove duplicated features after grouping step}
\usage{
removeduplicatedpeaks(msbatch, ppm, dmz = 5, drt = 30, thr_overlap = 0.7)
}
\arguments{
\item{msbatch}{msbatch object}

\item{ppm}{mz tolerance in ppm}
}
\value{
msbatch
}
\description{
Remove duplicated features after grouping step
}
\author{
M Isabel Alcoriza-Balaguer <maialba@iislafe.es>
}
\keyword{internal}
