% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalProcessing.R
\name{rtcorrection}
\alias{rtcorrection}
\title{Correct RT based on a rtmodel.}
\usage{
rtcorrection(rt, rtmodel)
}
\arguments{
\item{rt}{rt vector to correct.}

\item{rtmodel}{rt model.}
}
\value{
corrected rt vector.
}
\description{
Correct RT based on a rtmodel.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@iislafe.es>
}
\keyword{internal}
