% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalFunctions.R
\name{searchIsotopes}
\alias{searchIsotopes}
\title{Target isotopes search}
\usage{
searchIsotopes(
  msobject,
  label,
  adductsTable = LipidMS::adductsTable,
  rttol = 10,
  ppm = 10,
  coelCutoff = 0.8
)
}
\arguments{
\item{msobject}{a msobject generated by any of the identification functions.}

\item{label}{isotope employed for the experiment. It can be "13C" or "D".}

\item{adductsTable}{adducts table employed for lipids annotation.}

\item{rttol}{rt window in seconds.}

\item{ppm}{mass error tolerance.}

\item{coelCutoff}{coelution score threshold between isotopes. By default, 0.8.}
}
\value{
List with the isotopes for each compound in the results data frame.
}
\description{
This function uses annotation results of an unlabelled sample to search
for labelled compounds in a labelled sample.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
