% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcessing.R
\name{setmsbatch}
\alias{setmsbatch}
\title{Create msbatch for batch processing.}
\usage{
setmsbatch(msobjectlist, metadata)
}
\arguments{
\item{msobjectlist}{list of msobjects.}

\item{metadata}{sample metadata. Optional. It can be a csv file or a data.frame 
with 3 columns (sample, acquistionmode and sampletype).}
}
\value{
msbatch
}
\description{
Create msbatch from a list of msobjects to build an msbatch.
}
\details{
samples are sorted following the metadata data.frame.
}
\examples{
\dontrun{
msbatch <- setmsbatch(msobjectlist)
}

}
\seealso{
\link{dataProcessing} and \link{batchdataProcessing}
}
\author{
M Isabel Alcoriza-Balaguer <maialba@iislafe.es>
}
