% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sumChains}
\alias{sumChains}
\title{Calculate total number of carbons and double bounds of lipid chains}
\usage{
sumChains(chains, n)
}
\arguments{
\item{chains}{character value with the configuration of the chains separated
by a white space}

\item{n}{number of chains}
}
\value{
Character value indicating the total number of carbons and double
bounds
}
\description{
Given the structure of a lipid specie, it sums up the chains.
}
\author{
M Isabel Alcoriza-Balaguer <maialba@alumni.uv.es>
}
\keyword{internal}
